Fix compiler warnings

+++ server/src/main/scala/de/frosner/broccoli/services/InstanceService.scala
@@ -51,7 +51,7 @@ class InstanceService @Inject()(nomadClient: NomadClient,
     def run() =
       nomadService.requestStatuses(instances.values.map(_.id).toSet)
   }
-  private val scheduledTask = scheduler.scheduleAtFixedRate(task, 0, pollingFrequencySeconds, TimeUnit.SECONDS)
+  private val scheduledTask = scheduler.scheduleAtFixedRate(task, 0L, pollingFrequencySeconds.toLong, TimeUnit.SECONDS)
 
   sys.addShutdownHook {
     scheduledTask.cancel(false)
